requirements.txt
streamlit
ffmpeg
openai-whisper
torch
